+++ plugin-dev/Source/Sentry/Private/SentrySettings.cpp
@@ -2,14 +2,10 @@
 
 #include "SentrySettings.h"
 
-USentrySettings::USentrySettings(const FObjectInitializer& ObjectInitializer)
-	: Super(ObjectInitializer)
-	, InitAutomatically(false)
-	, UploadSymbolsAutomatically(false)
-{
-	DsnUrl = TEXT("");
-	Release = TEXT("");
-	ProjectName = TEXT("");
-	OrganisationName = TEXT("");
-	AuthToken = TEXT("");
+USentrySettings::USentrySettings(const FObjectInitializer &ObjectInitializer)
+    : Super(ObjectInitializer), InitAutomatically(false),
+      UploadSymbolsAutomatically(false) {
+  DsnUrl = TEXT("");
+  Release = TEXT("");
+  PropertiesFilePath = TEXT("Config/sentry.properties");
 }