+++ scripts/init-win.ps1
@@ -39,3 +39,4 @@ function setupPluginLink([string] $projectDir)
 
 setupPluginLink "sample"
 . "$PSScriptRoot/download-sdks.ps1"
+. "$PSScriptRoot/download-cli.ps1"