+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+- Add debug symbols upload ([#59](https://github.com/getsentry/sentry-unreal/pull/59))
 - Bump Cocoa SDK (iOS) to v7.21.0 ([#37](https://github.com/getsentry/sentry-unreal/pull/37))
   - [changelog](https://github.com/getsentry/sentry-cocoa/blob/master/CHANGELOG.md#7210)
   - [diff](https://github.com/getsentry/sentry-cocoa/compare/7.14.0...7.21.0)